Update README.md
# Nearest_Neighbor_Search_via_Feature_Extraction_Using_CNN

The goal of this project is to find commonalities between various photos using CNN models and feature extraction techniques.


<p align="justify">
With picture feature extraction, this research seeks to create a similarity search system that is effective. By utilising different CNN models, including VGG16, ResNet101, ZFNet, and MobileNet, we investigate feature extraction methods from a particular dataset. The method quickly finds the ten closest images based on feature similarity from an input image, effectively enabling content-based image retrieval.
Here, we contrast the dataset as a whole with the feature representation of a query image. To quickly calculate similarity scores, we use distance measures like the Euclidean distance and cosine similarity. The top 10 photographs that are the most comparable are then shown as the outcome.

</p>

## Output Sample 1:
![Output 1](https://www.linkpicture.com/q/Screenshot-2023-06-11-165648.jpg)

## Output Sample 2:
![Output 2](https://www.linkpicture.com/q/Screenshot-2023-06-11-170009.jpg)